amqqueue, amqqueue_v1: Fix type specs for Mnesia patterns
The solution is to move types specs out of the record itself to a
`-type()` (or more exactly an `-opaque()`) directive.

Consumers of this API are using the `amqqueue()` type anyway, so they
will have the proper type spec. And it allows us to define a second type
from the record, dedicated to the Mnesia pattern.